Enable LdapAuthentication to write data back to the ldap server
The new user cn=mwldapwriter,ou=users,dc=go2tech,dc=de is allowed to write
to the ou=users,dc=go2tech,dc=de dn subtree. This change configures LdapAuthentication
to use this user to change fields of the user from within MediaWiki (e.g. the password,
the e-mail address or the full name).

The same user needs to be used for searching of users, too, as reading anything from
ldap without any authentication is not permitted anymore.

This is the first step to remove go2tech.de/pwm (which currently is used for user
self-management).

Change-Id: Id806a6e96780e14fb6d6f0db04ce4d5771c9df4c