Allow non-operation members access to specific areas of opswiki
Opswiki will be the wiki, where ldap users can change settings, like their e-mail
address or their password. For this, a new group "Operations" is added, where all
current users in the ldap group "opswiki" are added. Some pages, such as Spezial:
Einstellungen, are now whitelisted to be read by anyone. Read permission is restricted
to Operations members, only (no user and no autoconfirmed). Also, autocreateaccount
is no permitted for * to allow the creation of user accounts when a user logs in
the first time to opswiki with the ldap credentials.

This change does not affect any _current_ opswiki member, they'll have the same
permissions as before. But now, any opswiki user (member of the opswiki ldap group)
have access to opswiki.

This is the second step to remove go2tech.de/pwm as the self-management tool.

Change-Id: I3a82468e26af80dbbbce35c4c6f090794e4b8826