Add additional logging to redis/logstash
This will, in the future, replace the file based logging through
wfDebugFile and friends. For now, both targets are used for logging.

Currently, there's a problem with the configuration when Wikibase
modules are loaded in the mediawiki/core composer vendor directory,
like we did so far with the composer-merge plugin. To fix this, the
composer dependencies will be unbundled with this commit. Each extension
and mediawiki/core needs to load the coposer dependencies. This will
hopefully be fixed when T151135 is resolved, so we can load all depen-
dencies in mediawiki/core again.

Change-Id: Ib09d8968cb1e4965b1bb3f3adc836e281ba2d06e