Remove CyanogenModDev and replace Dereferer with AmazonPartnernet
CMDev is not used anymore in any wiki (should've been deleted a long time
ago already). Dereferer was a hacky solution in a single PHP file without
any versioning of changes and quality assurance. It's just a simple extension,
but is now replaced by an improved version, AmazonPartnernet, which is
tested, has it's own repository.

Change-Id: If5003388b5e46899781b766e753bc7c1e2d02348