Add wgMainStash for all wikis to get ChronologyProtecter working
The ChronologyProtecter, which should make sure that the ordering of events are
consistent even when replicating data to replicas, needs an object storage that
can be used from any "datacenter" and keeps itself synchronized when multiple
instances are used. This property was not set so far for DroidWiki wikis resulting
in massive blocking sleep calls while the ChronologyProtecter falsly waits for the
replicas to catch up to the master.

Setting the main stash instance to redis, which is already configured, should solve
this problem. The configured redis instance is not local and can be reached from any
app server.

Fixes #1

Change-Id: I0cbd0f699b8641220610a2e78696ea478480dbcc